fix: 修改加锁逻辑

+++ polaris-domain/src/main/java/com/achobeta/domain/login/service/login/extendbiz/VerifyCodePostProcessor.java
@@ -32,21 +32,24 @@ public class VerifyCodePostProcessor implements LoginPostProcessor {
         String phone = tokenVO.getPhone();
         String code = tokenVO.getCode();
         log.info("正在校验手机号为{}的用户的验证码:{}",phone,code);
-        //加锁
-        operateCodeRepository.lockCheckCode(phone,code);
-        String codeByPhone = operateCodeRepository.getCodeByPhone(phone);
-        if(null == codeByPhone){
-            log.info("手机号为{}的用户的验证码不存在",phone);
-            throw new AppException(String.valueOf(GlobalServiceStatusCode.USER_CAPTCHA_CODE_ERROR.getCode()), GlobalServiceStatusCode.USER_CAPTCHA_CODE_ERROR.getMessage());
-        }else if(!codeByPhone.equals(code)){
-            log.info("手机号为{}的用户的验证码错误",phone);
-            throw new AppException(String.valueOf(GlobalServiceStatusCode.USER_CAPTCHA_CODE_ERROR.getCode()), GlobalServiceStatusCode.USER_CAPTCHA_CODE_ERROR.getMessage());
-        }else{
-            log.info("手机号为{}的用户的验证码正确",phone);
-            operateCodeRepository.deleteCode(phone,code);
+
+        try {
+            //加锁
+            operateCodeRepository.lockCheckCode(phone,code);
+            String codeByPhone = operateCodeRepository.getCodeByPhone(phone);
+            if(codeByPhone != null && codeByPhone.equals(code)){
+                log.info("手机号为{}的用户的验证码正确",phone);
+                operateCodeRepository.deleteCode(phone,code);
+
+                return true;
+            }
+            else{
+                log.info("手机号为{}的用户的验证码错误",phone);
+                throw new AppException(String.valueOf(GlobalServiceStatusCode.USER_CAPTCHA_CODE_ERROR.getCode()), GlobalServiceStatusCode.USER_CAPTCHA_CODE_ERROR.getMessage());
+            }
+        } finally {
             //解锁
-            operateCodeRepository.unlockCheckCode(phone,code);
-            return true;
+            operateCodeRepository.unlockCheckCode(phone, code);
         }
     }
